chore: updated page structure
