+++ .changeset/tasty-mails-rule.md
@@ -1,0 +1,5 @@
+---
+"copipe": minor
+---
+
+Updated page structure.
