+++ src/app/(side-menu-layout)/layout.tsx
@@ -1,0 +1,10 @@
+import SideMenuLayout from "@/modules/layouts/sideMenuLayout"
+import { ReactNode } from "react"
+
+export default function Layout({
+  children,
+}: {
+  children: ReactNode
+}) {
+  return <SideMenuLayout>{children}</SideMenuLayout>
+}
